% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_parse_single.R
\name{moc_parse_date}
\alias{moc_parse_date}
\title{Parse Date Field}
\usage{
moc_parse_date(x)
}
\arguments{
\item{x}{Character date string.}
}
\value{
Date or NA.
}
\description{
Normalizes date strings (allowing separators '-' or '/') to Date
  objects.
}
