% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_parse_single.R
\name{moc_parse_latlon}
\alias{moc_parse_latlon}
\title{Parse Latitude or Longitude String}
\usage{
moc_parse_latlon(x)
}
\arguments{
\item{x}{Character vector (first element used) representing a latitude or
longitude.}
}
\value{
Numeric decimal degrees or NA.
}
\description{
Converts a variety of latitude/longitude textual formats
  (degrees & minutes, decimal, hemisphere letters) into decimal degrees.
}
