% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_parse_single.R
\name{moc_read_file}
\alias{moc_read_file}
\title{Read Field Sheet File}
\usage{
moc_read_file(path)
}
\arguments{
\item{path}{Path to a tow JSON/TXT file.}
}
\value{
List with elements `path`, `raw`, `data`, and logical `ok`.
}
\description{
Reads a file and parses JSON content into a structured list.
}
