% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_parse_single.R
\name{moc_split_range}
\alias{moc_split_range}
\title{Split Time Range}
\usage{
moc_split_range(x)
}
\arguments{
\item{x}{Character, list, or vector representing a time range.}
}
\value{
Character vector of length 2: start time and end time (may contain
  NA).
}
\description{
Parses a time range expressed either as a single string (e.g.
  "12:00 - 12:30") or a list with start/end components into a length-2
  character vector.
}
