% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_parse_single.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null-Coalescing Helper}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value.}

\item{b}{Fallback value.}
}
\value{
Either `a` or `b`.
}
\description{
Returns the first argument if it is non-null and length > 0,
  otherwise the second.
}
